Move metadata put/get back to updateset
